+++ src/main/java/net/farkas/wildaside/particle/custom/StillSubstiliumParticle.java
@@ -20,7 +20,7 @@ public class StillSubstiliumParticle extends TextureSheetParticle {
     protected StillSubstiliumParticle(ClientLevel world, double x, double y, double z, double vx, double vy, double vz, SpriteSet spriteSet) {
         super(world, x, y, z);
         this.setSize(0.4f, 0.4f);
-        this.lifetime = (int)Math.max(1, 20 + (this.random.nextInt(20) - 10));
+        this.lifetime = Math.max(1, 20 + (this.random.nextInt(20) - 10));
         this.gravity = 0f;
         this.hasPhysics = false;
         this.xd = vx * 1;
